requirements.txt
boto3==1.7.42
botocore==1.10.42
docutils==0.14
ExifRead==2.1.2
futures==3.2.0
jmespath==0.9.3
psycopg2==2.7.5
python-dateutil==2.6.1
python-magic==0.4.13
PyYAML==3.12
s3cmd==2.0.0
s3transfer==0.1.13
six==1.10.0